ML-Training/requirements.txt
xgboost
scikit-learn
pandas
numpy
pg8000
joblib
python-dotenv
